requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asttokens==3.0.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
comm==0.2.3
cryptography==46.0.4
debugpy==1.8.20
decorator==5.2.1
distro==1.9.0
executing==2.2.1
google-auth==2.48.0
google-genai==1.62.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ipykernel==7.2.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jiter==0.13.0
jupyter_client==8.8.0
jupyter_core==5.9.1
matplotlib-inline==0.2.1
nest-asyncio==1.6.0
numpy==2.4.2
openai==2.17.0
packaging==26.0
pandas==3.0.0
parso==0.8.6
pexpect==4.9.0
platformdirs==4.5.1
prompt_toolkit==3.0.52
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pyzmq==27.1.0
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
tenacity==9.1.4
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
mcp>=1.0.0
openmeteo-requests>=0.3.0
requests-cache>=1.0.0
retry-requests>=2.0.0
uvicorn>=0.42.0
wcwidth==0.6.0
websockets==15.0.1
